requirements.txt
Django==4.1
pycryptodome==3.15.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
requests==2.28.1
django-filter==22.1
psycopg2==2.9.4
pynose==1.4.8
coverage==6.5.0
jsonnet==0.18.0
django-nose==1.4.6
django-rest-swagger==2.2.0
selenium==4.7.2
pynose==1.4.8
dj-database-url==2.1.0
whitenoise==6.5.0
gunicorn==21.2.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
openpyxl==3.1.2
redis==5.0.1
celery==5.3.6
Sphinx==6.2.1
sphinx-book-theme==1.0.1 
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
Jinja2==3.1.2
Pygments==2.17.2
docutils==0.19
snowballstemmer==2.2.0
Babel==2.14.0
alabaster==0.7.13
imagesize==1.4.1
packaging==23.2
MarkupSafe==2.1.3
charset-normalizer==2.1.1
idna==3.5
urllib3==1.26.18
certifi==2023.11.17
pydata-sphinx-theme==0.14.4
beautifulsoup4==4.12.2
Pygments==2.17.2
accessible-pygments==0.0.4
typing_extensions==4.8.0
soupsieve==2.5